Refactor large files into multiple smaller files under 700 lines
