+++ agent_plan.md
@@ -218,3 +218,6 @@
 * [2026-05-12] - Implemented Performance Fixes for CI and TS Types: Cleaned up mismatched properties between `SamplerBankParams` UI and internal type definitions (`coarseTune`, `fineTune`, `formantShift`, `quality`, `lockToSequencer`). Cleaned up custom sequencer optimization patch.
 * [2026-08-02] - Implemented Granular Pitch Shifter: Added `granularPitchShift` to `Note` and `SamplerBankParams` interfaces. Mapped parameters to `rubberband-processor.ts` via `SingingVoice.ts` and `useAudioEngine.ts` to allow independent granular pitch transposition logic via ratio scaling, decoupled from sample duration, fulfilling the "Granular Pitch Shifter" Innovation Lab idea. Added new idea: "Vocal Overdrive Worklet".
 * [2026-08-01] - Implemented Time-Stretch Envelope: Added `timeStretchEnvDepth` to `Note` and `SamplerBankParams` interfaces. Wired the parameter through `NoteSelector` and `SamplerPanel` UI down to `useAudioEngine.ts` and `SingingVoice.ts`, and finally applied the dynamic time stretch modulation via the `envelopeValue` inside `rubberband-processor.ts`. Fulfills the "Time-Stretch Envelope" Innovation Lab idea. Added new idea: "Granular Pitch Shifter".
+- Successfully refactored 10 out of 12 files that were over 1000 lines into multiple smaller files under 700 lines each.
+- Extracted structural UI repetition into sub-components for NoteSelector, RbsImportModal, AISongModal, and SamplerPanel.
+- Deferred splitting AISongStorage.ts and useAudioEngine.ts due to tightly coupled cyclic dependencies. These should be tackled incrementally in the future.
