+++ src/components/NoteSelector.tsx
@@ -1,3 +1,5 @@
+import { PropertyToggle } from './note-selector/PropertyToggle';
+import { PropertySlider } from "./note-selector/PropertySlider";
 import React, { memo } from 'react';
 import { NOTES, getScaleNotes } from '../utils/musicTheory';
 import type { ScaleDefinition } from '../utils/musicTheory';
@@ -179,110 +181,21 @@ export const NoteSelector: React.FC<NoteSelectorProps> = memo(({
                 </div>
 
                 {/* NEW: Duration Control */}
-                <fieldset className="flex flex-col gap-1 border-none p-0 m-0">
-                    <legend className="sr-only">Duration Control</legend>
-                    <div className="flex justify-between text-[10px] text-cyan-200/70 font-bold uppercase" aria-hidden="true">
-                        <label htmlFor="note-duration">Duration</label>
-                        <span className="text-cyan-400 font-mono text-[10px] drop-shadow-[0_0_5px_rgba(34,211,238,0.5)]">{currentLength} Steps</span>
-                    </div>
-                    <input
-                        id="note-duration"
-                        type="range"
-                        min="1"
-                        max="16"
-                        value={currentLength || 1}
-                        onChange={(e) => onLengthChange(parseInt(e.target.value))}
-                        className="w-full h-2 bg-gray-800 rounded-lg appearance-none cursor-pointer accent-cyan-400 border border-cyan-900/30 hover:accent-cyan-300 transition-all"
-                        aria-valuetext={`${currentLength} Steps`}
-                        aria-label="Duration"
-                    />
-                </fieldset>
+<PropertySlider label="Duration" id="note-duration" value={currentLength || 1} min={1} max={16} step={1} onChange={(v) => onLengthChange(v)} valueFormatter={(v) => `${v} Steps`} />
 
                 {onPropertyChange && (
                     <>
                         {/* Velocity Control */}
-                        <fieldset className="flex flex-col gap-1 border-none p-0 m-0">
-                            <legend className="sr-only">Velocity Control</legend>
-                            <div className="flex justify-between text-[10px] text-cyan-200/70 font-bold uppercase" aria-hidden="true">
-                                <label htmlFor="note-velocity">Velocity</label>
-                                <span className="text-emerald-400 font-mono text-[10px] drop-shadow-[0_0_5px_rgba(52,211,153,0.5)]">{Math.round((currentVelocity + 0.0001) * 100)}%</span>
-                            </div>
-                            <input
-                                id="note-velocity"
-                                type="range"
-                                min="0"
-                                max="1"
-                                step="0.01"
-                                value={currentVelocity}
-                                onChange={(e) => onPropertyChange('velocity', parseFloat(e.target.value))}
-                                className="w-full h-2 bg-gray-800 rounded-lg appearance-none cursor-pointer accent-emerald-400 border border-cyan-900/30 hover:accent-emerald-300 transition-all"
-                                aria-valuetext={`${Math.round((currentVelocity + 0.0001) * 100)}%`}
-                                aria-label="Velocity"
-                            />
-                        </fieldset>
+<PropertySlider label="Velocity" id="note-velocity" value={currentVelocity} onChange={(v) => onPropertyChange?.('velocity', v)} valueFormatter={(v) => `${Math.round(v * 100)}%`} ariaLabel="Velocity" />
 
                         {/* Timbre Control */}
-                        <fieldset className="flex flex-col gap-1 border-none p-0 m-0">
-                            <legend className="sr-only">Expression Control</legend>
-                            <div className="flex justify-between text-[10px] text-cyan-200/70 font-bold uppercase" aria-hidden="true">
-                                <label htmlFor="note-timbre">Expression</label>
-                                <span className="text-pink-400 font-mono text-[10px] drop-shadow-[0_0_5px_rgba(244,114,182,0.5)]">{Math.round((currentTimbre + 0.0001) * 100)}%</span>
-                            </div>
-                            <input
-                                id="note-timbre"
-                                type="range"
-                                min="0"
-                                max="1"
-                                step="0.01"
-                                value={currentTimbre}
-                                onChange={(e) => onPropertyChange('timbre', parseFloat(e.target.value))}
-                                className="w-full h-2 bg-gray-800 rounded-lg appearance-none cursor-pointer accent-pink-400 border border-pink-900/30 hover:accent-pink-300 transition-all"
-                                aria-valuetext={`${Math.round((currentTimbre + 0.0001) * 100)}%`}
-                                aria-label="Expression"
-                            />
-                        </fieldset>
+<PropertySlider label="Expression" id="note-timbre" value={currentTimbre} onChange={(v) => onPropertyChange?.('timbre', v)} valueFormatter={(v) => `${Math.round(v * 100)}%`} ariaLabel="Expression" />
 
                         {/* Probability Control */}
-                        <fieldset className="flex flex-col gap-1 border-none p-0 m-0">
-                            <legend className="sr-only">Probability Control</legend>
-                            <div className="flex justify-between text-[10px] text-cyan-200/70 font-bold uppercase" aria-hidden="true">
-                                <label htmlFor="note-prob">Probability</label>
-                                <span className="text-yellow-400 font-mono text-[10px] drop-shadow-[0_0_5px_rgba(250,204,21,0.5)]">{Math.round((currentProbability + 0.0001) * 100)}%</span>
-                            </div>
-                            <input
-                                id="note-prob"
-                                type="range"
-                                min="0"
-                                max="1"
-                                step="0.01"
-                                value={currentProbability}
-                                onChange={(e) => onPropertyChange('probability', parseFloat(e.target.value))}
-                                className="w-full h-2 bg-gray-800 rounded-lg appearance-none cursor-pointer accent-yellow-400 border border-yellow-900/30 hover:accent-yellow-300 transition-all"
-                                aria-valuetext={`${Math.round((currentProbability + 0.0001) * 100)}%`}
-                                aria-label="Probability"
-                            />
-                        </fieldset>
+<PropertySlider label="Probability" id="note-prob" value={currentProbability} onChange={(v) => onPropertyChange?.('probability', v)} valueFormatter={(v) => `${Math.round(v * 100)}%`} ariaLabel="Probability" />
 
                         {/* Microtiming Control */}
-                        <fieldset className="flex flex-col gap-1 border-none p-0 m-0">
-                            <legend className="sr-only">Microtiming Control</legend>
-                            <div className="flex justify-between text-[10px] text-cyan-200/70 font-bold uppercase" aria-hidden="true">
-                                <label htmlFor="note-micro">Microtiming</label>
-                                <span className="text-purple-400 font-mono text-[10px] drop-shadow-[0_0_5px_rgba(192,132,252,0.5)]">{currentMicrotiming > 0 ? '+' : ''}{currentMicrotiming.toFixed(2)}</span>
-                            </div>
-                            <input
-                                id="note-micro"
-                                type="range"
-                                min="-0.5"
-                                max="0.5"
-                                step="0.01"
-                                value={currentMicrotiming}
-                                onChange={(e) => onPropertyChange('microtiming', parseFloat(e.target.value))}
-                                className="w-full h-2 bg-gray-800 rounded-lg appearance-none cursor-pointer accent-purple-400 border border-purple-900/30 hover:accent-purple-300 transition-all"
-                                aria-valuetext={`${currentMicrotiming > 0 ? '+' : ''}${currentMicrotiming.toFixed(2)} steps`}
-                                aria-label="Microtiming"
-                            />
-                        </fieldset>
+<PropertySlider label="Microtiming" id="note-micro" value={currentMicrotiming} min={-0.5} max={0.5} onChange={(v) => onPropertyChange?.('microtiming', v)} valueFormatter={(v) => `${Math.round(v * 100)}%`} ariaLabel="Microtiming" />
 
                         {/* Freeze (Spectral Smear) Control */}
                         {trackType === "synth" && (
@@ -381,45 +294,9 @@ export const NoteSelector: React.FC<NoteSelectorProps> = memo(({
                                 />
                             </div>
 
-                            <div className="flex flex-col gap-1.5" role="group" aria-label="Bitcrush Override">
-                                <div className="flex justify-between items-center text-xs px-1">
-                                    <span className="text-gray-400 font-medium tracking-wide">Bitcrush</span>
-                                    <span className="text-indigo-400 font-mono bg-indigo-500/10 px-1.5 py-0.5 rounded shadow-inner border border-indigo-500/20">
-                                        {((currentBitcrush ?? 0) * 100).toFixed(0)}%
-                                    </span>
-                                </div>
-                                <input
-                                    type="range"
-                                    min="0"
-                                    max="1"
-                                    step="0.01"
-                                    value={currentBitcrush ?? 0}
-                                    onChange={(e) => onPropertyChange?.("bitcrush", parseFloat(e.target.value))}
-                                    className="w-full h-2 bg-gray-800 rounded-lg appearance-none cursor-pointer accent-indigo-400 border border-indigo-900/30 hover:accent-indigo-300 transition-all"
-                                    aria-valuetext={`${((currentBitcrush ?? 0) * 100).toFixed(0)} percent`}
-                                    aria-label="Bitcrush Amount"
-                                />
-                            </div>
+<PropertySlider label="Bitcrush" id="note-bitcrush" ariaLabel="Bitcrush Override" value={currentBitcrush ?? 0} onChange={(v) => onPropertyChange?.('bitcrush', v)} valueFormatter={() => `${((currentBitcrush ?? 0) * 100).toFixed(0)}%`} accentColor="accent-indigo-400 hover:accent-indigo-300" borderColor="border-indigo-900/30" />
 
-                            <div className="flex flex-col gap-1.5" role="group" aria-label="Downsample Override">
-                                <div className="flex justify-between items-center text-xs px-1">
-                                    <span className="text-gray-400 font-medium tracking-wide">Downsample</span>
-                                    <span className="text-indigo-400 font-mono bg-indigo-500/10 px-1.5 py-0.5 rounded shadow-inner border border-indigo-500/20">
-                                        {currentDownsample ?? 1}x
-                                    </span>
-                                </div>
-                                <input
-                                    type="range"
-                                    min="1"
-                                    max="32"
-                                    step="1"
-                                    value={currentDownsample ?? 1}
-                                    onChange={(e) => onPropertyChange?.("downsample", parseFloat(e.target.value))}
-                                    className="w-full h-2 bg-gray-800 rounded-lg appearance-none cursor-pointer accent-indigo-400 border border-indigo-900/30 hover:accent-indigo-300 transition-all"
-                                    aria-valuetext={`${currentDownsample ?? 1} times downsampled`}
-                                    aria-label="Downsample Factor"
-                                />
-                            </div>
+<PropertySlider label="Downsample" id="note-downsample" ariaLabel="Downsample Override" min={1} max={32} step={1} value={currentDownsample ?? 1} onChange={(v) => onPropertyChange?.('downsample', v)} valueFormatter={() => `${currentDownsample ?? 1}x`} accentColor="accent-indigo-400 hover:accent-indigo-300" borderColor="border-indigo-900/30" />
 
                             <div className="flex flex-col gap-1">
                                 <div className="flex justify-between text-[10px] text-cyan-200/70 font-bold uppercase">
@@ -444,67 +321,15 @@ export const NoteSelector: React.FC<NoteSelectorProps> = memo(({
 
 
                         {/* Pan Control */}
-                        <fieldset className="flex items-center justify-between group">
-                            <legend className="sr-only">Pan Control</legend>
-                            <div className="flex flex-col">
-                                <label htmlFor="note-pan" className="text-xs font-bold font-orbitron text-cyan-400 group-hover:text-cyan-300 transition-colors drop-shadow-[0_0_5px_rgba(6,182,212,0.3)]">Pan</label>
-                                <span className="text-cyan-400 font-mono text-[10px] drop-shadow-[0_0_5px_rgba(34,211,238,0.5)]">{currentPan !== undefined ? Math.round(currentPan * 100) : 0}%</span>
-                            </div>
-                            <input
-                                id="note-pan"
-                                type="range"
-                                min="-1"
-                                max="1"
-                                step="0.05"
-                                value={currentPan !== undefined ? currentPan : 0}
-                                onChange={(e) => onPropertyChange?.('pan', parseFloat(e.target.value))}
-                                className="w-24 h-1.5 bg-zinc-800 rounded-lg appearance-none cursor-pointer accent-cyan-500 hover:accent-cyan-400 focus:outline-none focus-visible:ring-2 focus-visible:ring-cyan-500 focus-visible:ring-offset-2 focus-visible:ring-offset-[#111827] shadow-[inset_0_1px_3px_rgba(0,0,0,0.5)]"
-                                aria-label="Pan"
-                            />
-                        </fieldset>
+<PropertySlider label="Pan" id="note-pan" value={currentPan || 0} min={-1} max={1} onChange={(v) => onPropertyChange?.('pan', v)} valueFormatter={(v) => v === 0 ? 'C' : (v < 0 ? `L${Math.round((v + 1) * 50)}` : `R${Math.round(v * 50)}`)} ariaLabel="Pan" />
 
 
                         {/* Pan Control */}
-                        <fieldset className="flex items-center justify-between group">
-                            <legend className="sr-only">Pan Control</legend>
-                            <div className="flex flex-col">
-                                <label htmlFor="note-pan" className="text-xs font-bold font-orbitron text-cyan-400 group-hover:text-cyan-300 transition-colors drop-shadow-[0_0_5px_rgba(6,182,212,0.3)]">Pan</label>
-                                <span className="text-cyan-400 font-mono text-[10px] drop-shadow-[0_0_5px_rgba(34,211,238,0.5)]">{currentPan !== undefined ? Math.round(currentPan * 100) : 0}%</span>
-                            </div>
-                            <input
-                                id="note-pan"
-                                type="range"
-                                min="-1"
-                                max="1"
-                                step="0.05"
-                                value={currentPan !== undefined ? currentPan : 0}
-                                onChange={(e) => onPropertyChange?.('pan', parseFloat(e.target.value))}
-                                className="w-24 h-1.5 bg-zinc-800 rounded-lg appearance-none cursor-pointer accent-cyan-500 hover:accent-cyan-400 focus:outline-none focus-visible:ring-2 focus-visible:ring-cyan-500 focus-visible:ring-offset-2 focus-visible:ring-offset-[#111827] shadow-[inset_0_1px_3px_rgba(0,0,0,0.5)]"
-                                aria-label="Pan"
-                            />
-                        </fieldset>
+<PropertySlider label="Pan" id="note-pan" value={currentPan || 0} min={-1} max={1} onChange={(v) => onPropertyChange?.('pan', v)} valueFormatter={(v) => v === 0 ? 'C' : (v < 0 ? `L${Math.round((v + 1) * 50)}` : `R${Math.round(v * 50)}`)} ariaLabel="Pan" />
 
                                 {/* Delay Send Control */}
                         {onPropertyChange && (
-                            <fieldset className="flex flex-col gap-1 border-none p-0 m-0">
-                                <legend className="sr-only">Delay Send Control</legend>
-                                <div className="flex justify-between text-[10px] text-cyan-200/70 font-bold uppercase" aria-hidden="true">
-                                    <label htmlFor="note-delay-send">Delay Send</label>
-                                    <span className="text-indigo-400 font-mono text-[10px] drop-shadow-[0_0_5px_rgba(129,140,248,0.5)]">{currentDelaySend !== undefined ? Math.round(currentDelaySend * 100) : 0}%</span>
-                                </div>
-                                <input
-                                    id="note-delay-send"
-                                    type="range"
-                                    min="0"
-                                    max="1"
-                                    step="0.01"
-                                    value={currentDelaySend !== undefined ? currentDelaySend : 0}
-                                    onChange={(e) => onPropertyChange('delaySend', parseFloat(e.target.value))}
-                                    className="w-full h-2 bg-gray-800 rounded-lg appearance-none cursor-pointer accent-indigo-400 border border-indigo-900/30 hover:accent-indigo-300 transition-all"
-                                    aria-valuetext={`${currentDelaySend !== undefined ? Math.round(currentDelaySend * 100) : 0}%`}
-                                    aria-label="Delay Send"
-                                />
-                            </fieldset>
+<PropertySlider label="Delay Send" id="note-delay-send" value={currentDelaySend !== undefined ? currentDelaySend : 0} onChange={(v) => onPropertyChange?.('delaySend', v)} valueFormatter={() => currentDelaySend !== undefined ? `${Math.round(currentDelaySend * 100)}%` : 'OFF'} ariaLabel="Delay Send Control" />
                         )}
                         {/* Rhythmic Gate Parameters */}
                         {trackType === 'synth' && (   // or 'sampler' if you want both
@@ -901,67 +726,13 @@ export const NoteSelector: React.FC<NoteSelectorProps> = memo(({
                         )}
 
                         {/* Filter Cutoff Control */}
-                        <fieldset className="flex flex-col gap-1 border-none p-0 m-0">
-                            <legend className="sr-only">Filter Cutoff Control</legend>
-                            <div className="flex justify-between text-[10px] text-cyan-200/70 font-bold uppercase" aria-hidden="true">
-                                <label htmlFor="note-cutoff">Filter Cutoff</label>
-                                <span className="text-cyan-400 font-mono text-[10px] drop-shadow-[0_0_5px_rgba(34,211,238,0.5)]">{currentFilterCutoff !== undefined ? Math.round(currentFilterCutoff * 100) : 100}%</span>
-                            </div>
-                            <input
-                                id="note-cutoff"
-                                type="range"
-                                min="0"
-                                max="1"
-                                step="0.01"
-                                value={currentFilterCutoff !== undefined ? currentFilterCutoff : 1.0}
-                                onChange={(e) => onPropertyChange('filterCutoff', parseFloat(e.target.value))}
-                                className="w-full h-2 bg-gray-800 rounded-lg appearance-none cursor-pointer accent-cyan-400 border border-cyan-900/30 hover:accent-cyan-300 transition-all"
-                                aria-valuetext={`${currentFilterCutoff !== undefined ? Math.round(currentFilterCutoff * 100) : 100}%`}
-                                aria-label="Filter Cutoff"
-                            />
-                        </fieldset>
+<PropertySlider label="Filter Cutoff" id="note-cutoff" value={currentFilterCutoff !== undefined ? currentFilterCutoff : 1} onChange={(v) => onPropertyChange?.('filterCutoff', v)} valueFormatter={() => currentFilterCutoff !== undefined ? `${Math.round(currentFilterCutoff * 100)}%` : 'OFF'} ariaLabel="Filter Cutoff Override" />
 
                         {/* Filter Resonance Control */}
-                        <fieldset className="flex flex-col gap-1 border-none p-0 m-0">
-                            <legend className="sr-only">Filter Resonance Control</legend>
-                            <div className="flex justify-between text-[10px] text-cyan-200/70 font-bold uppercase" aria-hidden="true">
-                                <label htmlFor="note-resonance">Filter Res</label>
-                                <span className="text-cyan-400 font-mono text-[10px] drop-shadow-[0_0_5px_rgba(34,211,238,0.5)]">{currentFilterResonance !== undefined ? Math.round(currentFilterResonance * 100) : 0}%</span>
-                            </div>
-                            <input
-                                id="note-resonance"
-                                type="range"
-                                min="0"
-                                max="1"
-                                step="0.01"
-                                value={currentFilterResonance || 0}
-                                onChange={(e) => onPropertyChange('filterResonance', parseFloat(e.target.value))}
-                                className="w-full h-2 bg-gray-800 rounded-lg appearance-none cursor-pointer accent-cyan-400 border border-cyan-900/30 hover:accent-cyan-300 transition-all"
-                                aria-valuetext={`${currentFilterResonance !== undefined ? Math.round(currentFilterResonance * 100) : 0}%`}
-                                aria-label="Filter Resonance"
-                            />
-                        </fieldset>
+<PropertySlider label="Filter Res" id="note-resonance" value={currentFilterResonance !== undefined ? currentFilterResonance : 0} onChange={(v) => onPropertyChange?.('filterResonance', v)} valueFormatter={() => currentFilterResonance !== undefined ? `${Math.round(currentFilterResonance * 100)}%` : 'OFF'} ariaLabel="Filter Resonance Override" />
 
                         {/* Envelope Mod Control */}
-                        <fieldset className="flex flex-col gap-1 border-none p-0 m-0">
-                            <legend className="sr-only">Envelope Modulation Control</legend>
-                            <div className="flex justify-between text-[10px] text-slate-400 font-bold uppercase" aria-hidden="true">
-                                <label htmlFor="note-envmod">Env Mod</label>
-                                <span className="text-emerald-400">{currentEnvMod !== undefined ? Math.round(currentEnvMod * 100) : 50}%</span>
-                            </div>
-                            <input
-                                id="note-envmod"
-                                type="range"
-                                min="0"
-                                max="1"
-                                step="0.01"
-                                value={currentEnvMod !== undefined ? currentEnvMod : 0.5}
-                                onChange={(e) => onPropertyChange('envMod', parseFloat(e.target.value))}
-                                className="w-full h-2 bg-slate-700 rounded-lg appearance-none cursor-pointer accent-emerald-500"
-                                aria-valuetext={`${currentEnvMod !== undefined ? Math.round(currentEnvMod * 100) : 50}%`}
-                                aria-label="Envelope Modulation"
-                            />
-                        </fieldset>
+<PropertySlider label="Env Mod" id="note-envmod" min={-1} max={1} step={0.01} value={currentEnvMod !== undefined ? currentEnvMod : 0} onChange={(v) => onPropertyChange?.('envMod', v)} valueFormatter={() => currentEnvMod !== undefined ? `${currentEnvMod > 0 ? '+' : ''}${Math.round(currentEnvMod * 100)}%` : 'OFF'} ariaLabel="Env Mod Override" />
 
                         {/* Retrigger (Ratchet) Control */}
                         <fieldset className="flex flex-col gap-1 pb-1 border-none p-0 m-0">
