+++ src/engines/rubberband/performance/PerformanceOptimizer.ts
@@ -1,188 +1,10 @@
-/**
- * PerformanceOptimizer - Real-time performance optimizations for Rubber Band
- *
- * Part of RUBBERBAND_ENHANCEMENT_PLAN Section 7: Real-Time Performance Optimizations
- *
- * Provides:
- * - Pre-allocated circular buffers in WASM memory
- * - SIMD instruction support detection
- * - Adaptive quality based on CPU load
- * - Worker pool for parallel TTS generation
- *
- * @see RUBBERBAND_ENHANCEMENT_PLAN.md Section 7
- */
+import type {
+  CpuLoadStatus, PerformanceConfig, WasmMemoryAllocation, CircularBuffer, QualityLevel, QualityConfig, WorkerStatus, WorkerTask, TTSWorkerMessage, WASMModule
+} from './types';
+import { DEFAULT_PERFORMANCE_CONFIG, QUALITY_CONFIGS } from './types';
 
-/** CPU load monitoring result */
-export interface CpuLoadStatus {
-    /** Current CPU load estimate (0.0 - 1.0) */
-    load: number;
-    /** Whether system is overloaded */
-    isOverloaded: boolean;
-    /** Recommended quality setting */
-    recommendedQuality: 'high' | 'medium' | 'low';
-    /** Estimated processing headroom in ms */
-    headroomMs: number;
-}
+import { createTTSWorkerBlob } from './worker';
 
-/** Configuration for performance optimizer */
-export interface PerformanceConfig {
-    /** Buffer size for WASM processing (default: 4096) */
-    bufferSize: number;
-    /** Number of worker threads for TTS (default: 4) */
-    workerPoolSize: number;
-    /** CPU load threshold for quality reduction (default: 0.8) */
-    overloadThreshold: number;
-    /** Enable SIMD optimizations if available (default: true) */
-    enableSimd: boolean;
-    /** Enable SharedArrayBuffer for lock-free communication (default: true) */
-    enableSharedArrayBuffer: boolean;
-    /** Sample rate for audio processing (default: 44100) */
-    sampleRate: number;
-    /** Number of channels (default: 2 for stereo) */
-    numChannels: number;
-    /** Enable adaptive buffer sizing based on load (default: true) */
-    adaptiveBufferSizing: boolean;
-    /** Polling interval for CPU monitoring in ms (default: 100) */
-    cpuMonitorIntervalMs: number;
-}
-
-/** Default performance configuration */
-export const DEFAULT_PERFORMANCE_CONFIG: PerformanceConfig = {
-    bufferSize: 4096,
-    workerPoolSize: 4,
-    overloadThreshold: 0.8,
-    enableSimd: true,
-    enableSharedArrayBuffer: true,
-    sampleRate: 44100,
-    numChannels: 2,
-    adaptiveBufferSizing: true,
-    cpuMonitorIntervalMs: 100
-};
-
-/** WASM memory allocation info */
-export interface WasmMemoryAllocation {
-    /** Pointer to input buffer in WASM memory */
-    inputPtr: number;
-    /** Pointer to output buffer in WASM memory */
-    outputPtr: number;
-    /** Buffer size in samples */
-    bufferSize: number;
-    /** Number of channels */
-    numChannels: number;
-    /** Whether allocation was successful */
-    success: boolean;
-    /** Error message if allocation failed */
-    error?: string;
-}
-
-/** Circular buffer for lock-free audio processing */
-export interface CircularBuffer {
-    /** SharedArrayBuffer for lock-free access */
-    buffer: SharedArrayBuffer;
-    /** Typed array view for convenient access */
-    view: Float32Array;
-    /** Current write position (Atomic) */
-    writeIndex: Int32Array;
-    /** Current read position (Atomic) */
-    readIndex: Int32Array;
-    /** Buffer capacity in samples */
-    capacity: number;
-    /** Number of channels */
-    numChannels: number;
-}
-
-/** Worker pool task */
-interface WorkerTask {
-    id: string;
-    type: 'tts' | 'process' | 'pitch';
-    data: unknown;
-    resolve: (value: unknown) => void;
-    reject: (reason: Error) => void;
-    timestamp: number;
-}
-
-/** Worker status */
-interface WorkerStatus {
-    worker: Worker;
-    busy: boolean;
-    taskId: string | null;
-    lastActivity: number;
-}
-
-/** Quality levels for adaptive processing */
-export type QualityLevel = 'high' | 'medium' | 'low';
-
-/** Quality configuration mapping */
-export interface QualityConfig {
-    /** Engine type: 'finer' for quality, 'faster' for speed */
-    engine: 'finer' | 'faster';
-    /** Transient handling: 'crisp', 'mixed', or 'smooth' */
-    transients: 'crisp' | 'mixed' | 'smooth';
-    /** Phase mode: 'laminar' or 'independent' */
-    phase: 'laminar' | 'independent';
-    /** Formant handling: 'preserved' or 'shifted' */
-    formant: 'preserved' | 'shifted';
-    /** Buffer size multiplier */
-    bufferMultiplier: number;
-    /** Enable SIMD processing */
-    useSimd: boolean;
-}
-
-/** Quality level configurations */
-export const QUALITY_CONFIGS: Record<QualityLevel, QualityConfig> = {
-    high: {
-        engine: 'finer',
-        transients: 'crisp',
-        phase: 'laminar',
-        formant: 'preserved',
-        bufferMultiplier: 1,
-        useSimd: true
-    },
-    medium: {
-        engine: 'faster',
-        transients: 'mixed',
-        phase: 'laminar',
-        formant: 'preserved',
-        bufferMultiplier: 1,
-        useSimd: true
-    },
-    low: {
-        engine: 'faster',
-        transients: 'smooth',
-        phase: 'independent',
-        formant: 'shifted',
-        bufferMultiplier: 2,
-        useSimd: false
-    }
-};
-
-/** TTS Worker message types */
-interface TTSWorkerMessage {
-    type: 'init' | 'generate' | 'result' | 'error' | 'ready';
-    id?: string;
-    data?: unknown;
-    error?: string;
-}
-
-/** WASM Module interface */
-interface WASMModule {
-    _malloc(size: number): number;
-    _free(ptr: number): void;
-    HEAPU8: Uint8Array;
-    HEAPF32: Float32Array;
-    simd?: boolean;
-}
-
-/**
- * PerformanceOptimizer class for maximizing real-time audio performance.
- *
- * Features:
- * - SIMD detection and utilization
- * - Pre-allocated WASM memory buffers
- * - Circular buffer management with SharedArrayBuffer
- * - TTS worker pool with load balancing
- * - Adaptive quality based on CPU load monitoring
- */
 export class PerformanceOptimizer {
     private config: PerformanceConfig;
     private workers: WorkerStatus[] = [];
@@ -966,206 +788,5 @@ export class PerformanceOptimizer {
  *
  * @returns Blob URL for the worker script
  */
-export function createTTSWorkerBlob(): string {
-    const workerCode = `
-        // TTS Worker for parallel text-to-speech generation
-        let synth = null;
-
-        // Initialize speech synthesis if available
-        if (typeof self !== 'undefined' && 'speechSynthesis' in self) {
-            synth = self.speechSynthesis;
-        }
-
-        self.onmessage = function(event) {
-            const { type, id, data } = event.data;
-
-            switch (type) {
-                case 'init':
-                    self.postMessage({ type: 'ready', id });
-                    break;
-
-                case 'tts':
-                case 'generate':
-                    handleTTS(id, data);
-                    break;
-
-                case 'process':
-                    handleProcess(id, data);
-                    break;
-
-                case 'pitch':
-                    handlePitchShift(id, data);
-                    break;
-
-                default:
-                    self.postMessage({
-                        type: 'error',
-                        id,
-                        error: 'Unknown task type: ' + type
-                    });
-            }
-        };
-
-        function handleTTS(id, data) {
-            try {
-                const { text, pitch = 1.0, rate = 1.0 } = data || {};
-
-                if (!synth) {
-                    // Fallback: generate placeholder audio data
-                    const sampleRate = 22050;
-                    const duration = (text?.length || 10) * 0.1; // Rough estimate
-                    const samples = Math.floor(sampleRate * duration);
-                    const audio = new Float32Array(samples);
-
-                    // Generate simple waveform as placeholder
-                    for (let i = 0; i < samples; i++) {
-                        const t = i / sampleRate;
-                        audio[i] = Math.sin(2 * Math.PI * 440 * pitch * t) * 0.1;
-                        // Add envelope
-                        if (t < 0.01) audio[i] *= t / 0.01;
-                        if (t > duration - 0.01) audio[i] *= (duration - t) / 0.01;
-                    }
-
-                    self.postMessage({
-                        type: 'result',
-                        id,
-                        data: { audio, sampleRate, text }
-                    }, [audio.buffer]);
-                    return;
-                }
-
-                // Use Web Speech API
-                const utterance = new SpeechSynthesisUtterance(text);
-                utterance.pitch = pitch;
-                utterance.rate = rate;
-
-                // Note: Actual audio buffer extraction from Web Speech API
-                // requires MediaRecorder or AudioWorklet capture on main thread
-                // This is a simplified implementation
-
-                utterance.onend = () => {
-                    self.postMessage({
-                        type: 'result',
-                        id,
-                        data: { text, pitch, rate, completed: true }
-                    });
-                };
-
-                utterance.onerror = (e) => {
-                    self.postMessage({
-                        type: 'error',
-                        id,
-                        error: e.error
-                    });
-                };
-
-                synth.speak(utterance);
-            } catch (error) {
-                self.postMessage({
-                    type: 'error',
-                    id,
-                    error: error.message
-                });
-            }
-        }
-
-        function handleProcess(id, data) {
-            try {
-                const { buffer, operation } = data || {};
-
-                if (!buffer || !(buffer instanceof Float32Array)) {
-                    throw new Error('Invalid buffer data');
-                }
-
-                const result = new Float32Array(buffer.length);
-
-                switch (operation) {
-                    case 'normalize':
-                        const max = Math.max(...buffer.map(Math.abs));
-                        const scale = max > 0 ? 1 / max : 1;
-                        for (let i = 0; i < buffer.length; i++) {
-                            result[i] = buffer[i] * scale;
-                        }
-                        break;
-
-                    case 'fadeIn':
-                        const fadeInLength = Math.min(buffer.length, Math.floor(0.01 * 44100));
-                        for (let i = 0; i < buffer.length; i++) {
-                            const gain = i < fadeInLength ? i / fadeInLength : 1;
-                            result[i] = buffer[i] * gain;
-                        }
-                        break;
-
-                    case 'fadeOut':
-                        const fadeOutLength = Math.min(buffer.length, Math.floor(0.01 * 44100));
-                        for (let i = 0; i < buffer.length; i++) {
-                            const gain = i >= buffer.length - fadeOutLength
-                                ? (buffer.length - i) / fadeOutLength
-                                : 1;
-                            result[i] = buffer[i] * gain;
-                        }
-                        break;
-
-                    default:
-                        result.set(buffer);
-                }
-
-                self.postMessage({
-                    type: 'result',
-                    id,
-                    data: { result, operation }
-                }, [result.buffer]);
-            } catch (error) {
-                self.postMessage({
-                    type: 'error',
-                    id,
-                    error: error.message
-                });
-            }
-        }
-
-        function handlePitchShift(id, data) {
-            try {
-                const { buffer, ratio, sampleRate = 44100 } = data || {};
-
-                if (!buffer || !(buffer instanceof Float32Array)) {
-                    throw new Error('Invalid buffer data');
-                }
-
-                // Simple time-domain pitch shift using resampling
-                // For production, use Rubber Band or similar
-                const outputLength = Math.floor(buffer.length / ratio);
-                const result = new Float32Array(outputLength);
-
-                for (let i = 0; i < outputLength; i++) {
-                    const sourceIndex = i * ratio;
-                    const index1 = Math.floor(sourceIndex);
-                    const index2 = Math.min(index1 + 1, buffer.length - 1);
-                    const frac = sourceIndex - index1;
-
-                    result[i] = buffer[index1] * (1 - frac) + buffer[index2] * frac;
-                }
-
-                self.postMessage({
-                    type: 'result',
-                    id,
-                    data: { result, ratio, sampleRate }
-                }, [result.buffer]);
-            } catch (error) {
-                self.postMessage({
-                    type: 'error',
-                    id,
-                    error: error.message
-                });
-            }
-        }
-
-        // Signal ready on load
-        self.postMessage({ type: 'ready' });
-    `;
-
-    const blob = new Blob([workerCode], { type: 'application/javascript' });
-    return URL.createObjectURL(blob);
-}
 
 export default PerformanceOptimizer;
