+++ src/engines/rubberband/performance/index.ts
@@ -1,0 +1,5 @@
+export * from './types';
+export * from './PerformanceOptimizer';
+export * from './worker';
+import PerformanceOptimizer from './PerformanceOptimizer';
+export default PerformanceOptimizer;
